Move credit usage ownership to sidebar user menu
